Remove location – add confirmation

+++ js/app.controller.js
@@ -69,6 +69,7 @@ function renderLocs(locs) {
 }
 
 function onRemoveLoc(locId) {
+    if(!confirm('are you sure?'))return
     locService.remove(locId)
         .then(() => {
             flashMsg('Location removed')
